+++ packages/logs/src/boot/startLogs.ts
@@ -1,5 +1,6 @@
 import type { Context, TelemetryEvent } from '@datadog/browser-core'
 import {
+  willSyntheticsInjectRum,
   areCookiesAuthorized,
   canUseEventBridge,
   getEventBridge,
@@ -20,6 +21,7 @@ import type { CommonContext } from '../rawLogsEvent.types'
 import { startLogsBatch } from '../transport/startLogsBatch'
 import { startLogsBridge } from '../transport/startLogsBridge'
 import type { Logger } from '../domain/logger'
+import { startInternalContext } from '../domain/internalContext'
 
 export function startLogs(configuration: LogsConfiguration, getCommonContext: () => CommonContext, mainLogger: Logger) {
   const lifeCycle = new LifeCycle()
@@ -47,7 +49,7 @@ export function startLogs(configuration: LogsConfiguration, getCommonContext: ()
   const { handleLog } = startLoggerCollection(lifeCycle)
 
   const session =
-    areCookiesAuthorized(configuration.cookieOptions) && !canUseEventBridge()
+    areCookiesAuthorized(configuration.cookieOptions) && !canUseEventBridge() && !willSyntheticsInjectRum()
       ? startLogsSessionManager(configuration)
       : startLogsSessionManagerStub(configuration)
 
@@ -59,8 +61,11 @@ export function startLogs(configuration: LogsConfiguration, getCommonContext: ()
     startLogsBridge(lifeCycle)
   }
 
+  const internalContext = startInternalContext(session)
+
   return {
     handleLog,
+    getInternalContext: internalContext.get,
   }
 }